remote development utiltilies
git-svn-id: http://www.ifountain.org/repos/os/RapidModules@7940 a08bcbf9-5f6c-4d33-af0a-e494bc2cacca

+++ RapidCMDB/devDocs/developerUtilities/remoteDevelopmentSolution/customerSideScripts/fileOperation.groovy
@@ -14,7 +14,17 @@ if(operation == "copy")
 }
 else if(operation == "delete")
 {
-    targetFile.delete();
+    if(targetFile.exists())
+    {
+        if(!targetFile.isDirectory())
+        {
+            targetFile.delete();
+        }
+        else
+        {
+            throw new Exception("You cannot delete a nonempty directory.");            
+        }
+    }
 }
 return "<successful>succesfully copied file</successful>"
